requirements.txt
django==4.0.2
Pillow==9.0.1
django-import-export
psycopg2
djangorestframework